Implemented optimizer and losses

+++ lib/rubyzero/nn/layers/affine.rb
@@ -1,9 +1,16 @@
 module RubyZero::NN::Layers
+    # Affine layer
     class Affine < Layer
+        # @param input_size [Integer] input size
+        # @param output_size [Integer] output size
+        # @param bias [Boolean] whether to use bias
         def initialize(input_size, output_size, bias: true)
-            @weight = RubyZero::Float64.randn([input_size, output_size])
-            @bias = RubyZero::Float64.randn([output_size]) if bias
+            @weight = RubyZero::Float32.randn([input_size, output_size])
+            @bias = RubyZero::Float32.randn([output_size]) if bias
         end
+        # Forward pass
+        # @param x [RubyZero::Float32] input
+        # @return [RubyZero::Float32] output
         def forward(x)
             x = x.dot(@weight)
             if @bias
